workflow101/requirements.txt
django==4.1.4
celery
django-guardian
djangorestframework
django-filter
django-formtools